display video thumbnails
