+++ scripts/utils/getMediasByPhotographer.js
@@ -5,34 +5,6 @@ export async function getMediasByPhotographer() {
   const params = (new URL(document.location).searchParams);
   const id = parseInt(params.get("id"));
 
-  // console.log(media.filter(media => media.photographerId === id))
   const medias =  media.filter(media => media.photographerId === id)
-  // console.log(medias);
-  const dropdown = document.getElementById("sort__by");
-  // console.log(dropdown);
   return medias
-}
-
-export async function sortMedias() {
-  const dropdown = document.getElementById("sort__by").value;
-  // console.log(dropdown);
-  const data = await getMediasByPhotographer();
-
-  switch (dropdown) {
-    case "Popularity":
-      data.sort(function(a, b){
-        return b.likes - a.likes
-      })
-      break;
-    case "Date":
-      data.sort(function(a, b){
-        return b.date - a.date
-      })
-      break;
-    case "Title":
-      data.sort(function(a, b){
-        return a.title - b.title
-      })
-      break;
-  }
 }
